Re-adds my content on a4.php file
I add my content on a4.php file. As I said on the canvas, I focused Value of Scrum instead of Benefits of Scrum because the Scrum Guide says about "Value of Scrum." I am going to edit the line related to this in index.html later, not today. Thanks,